(PC-7377) Refacto: move appstate to libs
